# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Application API for project WineApp",
        "version": "v1"
    },
    "paths": {
...
        "/api/v1/vineyardWine/{vineyardWineId}": {
            "get": {
                "tags": [
                    "VineyardWine"
                ],
                "summary": "Show vineyard by vineyardWineId",
                "description": "Show vineyard by vineyardWineId",
                "operationId": "showVineyardWineById",
                "parameters": [
                    {
                        "name": "vineyardWineId",
                        "in": "path",
                        "description": "VineyardWine ID",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Response Successfull",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            },
            "put": {
                "tags": [
                    "VineyardWine"
                ],
                "summary": "Update VineyardWine",
                "description": "Update VineyardWine by selected vineyardWineId",
                "operationId": "updateVineyardWine",
                "parameters": [
                    {
                        "name": "vineyardWineId",
                        "in": "path",
                        "description": "VineyardWine ID",
                        "required": true
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [],
                                "properties": {
                                    "vineyard_id": {
                                        "type": "integer"
                                    },
                                    "wine_id": {
                                        "type": "integer"
                                    },
                                    "title": {
                                        "type": "string"
                                    },
                                    "quantity": {
                                        "type": "integer"
                                    },
                                    "note": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Response Successfull",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/api/v1/vineyardWine": {
            "post": {
                "tags": [
                    "VineyardWine"
                ],
                "summary": "Create VineyardWine",
                "description": "Create VineyardWine",
                "operationId": "createVineyardWine",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "vineyard_id",
                                    "wine_id",
                                    "title",
                                    "quantity"
                                ],
                                "properties": {
                                    "vineyard_id": {
                                        "type": "integer"
                                    },
                                    "wine_id": {
                                        "type": "integer"
                                    },
                                    "title": {
                                        "type": "string"
                                    },
                                    "quantity": {
                                        "type": "integer"
                                    },
                                    "note": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Response Successfull",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/api/v1/vineyardWine/vineyard/{vineyardId}": {
            "get": {
                "tags": [
                    "VineyardWine"
                ],
                "summary": "Show vineyardWines by vineyardId",
                "description": "Show vineyardWines by vineyardId",
                "operationId": "showVineyardWinesByVineyardId",
                "parameters": [
                    {
                        "name": "vineyardId",
                        "in": "path",
                        "description": "Vineyard ID",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Response Successfull",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
...
    }
}